requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
plotly==5.24.1
streamlit==1.37.1
pyarrow==17.0.0
duckdb==1.0.0
pydantic==